requirements.txt
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
paramiko==2.7.2
Pillow==8.2.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
